requirements.txt
boto3
Flask
zappa
Pillow